[Site] Fix randomized IDs
In an effort to prevent ID clashes, some IDs are randomized (search
for "Math.random"). This worked well with `gatsby develop` but when
the ID was also referenced from JS, they broke with
`gatsby build && gatsby serve`. Apparently, the JS code did not
capture the once-randomized ID variable and instead rerolled the
dice on every execution. As a consequence, `document.getElementById`
never found the referenced elements because the IDs differed.

Instead of tracking this down to the root cause, I simply removed
the randomization because, luckily!, it was never needed in the
cases where IDs were referenced from JS. (This is different for
the remaining `Math.random` uses, which need some form to prevent
ID clashes.)