[Site] Limit large Twitter cards to pages with non-default images
The site uses a picture of me as default social media preview. On
large-immage summary cards, Twitter usually cuts the image to a 16:9
area, which results in my head not being in the picture. That looks
stupid.

With this change, pages that don't define their own picture, use the
default summary card, which Twitter usually displays with a square
image next to the page details. That includes my head. :)