[Site] Replace "//" with " - " in meta title #84
Google shows many search results for the new site as follows:

> Value-Based Classes // nipafx - Java 9

The "Java 9" makes no sense and it is unclear to me why it is there.
One possible reason is that Google expects site content and brand to
be separated by " - " and, when not finding that, looks for strong
related terms to apply as "brand", so for me it finds Java 9. Maybe?

Let's see...