[Site] Remove `rel="noopener noreferrer"` from external links
Initially, external links were opened in new tabs, which leads to
Lighthouse complaining about "unsafe cross-origin destinations" and
recommends to add `noopener` or `noreferrer`. Apparently, I've...

(a) not been reading that attentively and added both
(b) forgotten to remove the attributes when removing `target="_blank"`

This change remedies that by removing the attributes so that traffic
from my site is visible in other people's analytics.

[1] https://web.dev/external-anchors-use-rel-noopener/?utm_source=lighthouse&utm_medium=devtools