[Site] Remove `node_modules` from cache
Caching `node_module` is risky, e.g. because packages like
image_sharp use native code and need to be reinstalled when the node
version changes. This seems to be happening now after the update to
Node 17 (at least some times).

An alternative would be to include the node version in the cache's
name, but I don't want to fiddle with that now.