[Site] Update dependencies and remove bg-img and fontawesome-css
gatsby-background-image uses the outdated gatsby-plugin-image (as
opposed to the new gatsby-image) and its outdated fixed/fluid
approach and also requires Gatsby <= 4.

gatsby-plugin-fontawesome-css requires and old version of
@fortawesome/fontawesome-svg-core (if I correctly interpret the peer
dependency thingy), so I experimented with removing it. I checked
the FontAwesome icons in the header and share box and and didn't
notice any jumps, so I stuck with the removal.